requirements.txt
# Care-Tracker Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
python-multipart==0.0.18
jinja2==3.1.4
aiosqlite==0.20.0
pytz==2024.2
twilio==9.4.1
apscheduler==3.10.4
python-dotenv==1.0.1

# Testing
pytest==8.0.0
httpx==0.27.0